+++ amazonorders/entity/item.py
@@ -10,14 +10,14 @@ from amazonorders.entity.seller import Seller
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2024, Alex Laird"
-__version__ = "1.0.7"
+__version__ = "1.0.9"
 
 logger = logging.getLogger(__name__)
 
 
 class Item(Parsable):
     """
-    An Item in an Amazon `~amazonorders.entity.order.Order`.
+    An Item in an Amazon :class:`~amazonorders.entity.order.Order`.
     """
 
     def __init__(self,